Fix policy slice interface to value
For now we only support a single item return in dynamo. It was
possible to return a slice of data which was not correctly
converted to a term. Tests were passing because we only
had negative tests and the input for the test was not correctly
serialized to JSON.

I found this issue while trying to write getting started documentation
and writing out a use case.